refactor: merge patcher matchers, split page.rs, doc fixes (0.4.x)
- patcher: one build_matcher() helper replaces the two near-identical
  Aho-Corasick getters.
- page.rs (1815 -> 1416): move Element/BoundingBox to element.rs and the
  keyboard machinery to keyboard.rs. Pure relocation, public API unchanged
  (eoka::Element / eoka::page::Element still resolve).
- README: add connect-to-existing-Chrome section, correct the stale LOC claim.
